publish and unique-check test in a same file
